Move error tests to their own file
Move error tests to their own file
